requirements.txt
pyyaml>=6.0
jinja2>=3.1.0
watchdog>=3.0.0
click>=8.0.0
python-dotenv>=1.0.0
requests>=2.28.0
cryptography>=3.4.0 